refactor: Add "mode" param to seat component
